requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
Django==4.2.7
django-ckeditor==6.7.0
django-easy-audit==1.3.6
django-jazzmin==2.6.0
django-js-asset==2.1.0
django-shortuuidfield==0.1.3
django-taggit==5.0.1
Pillow==10.1.0
python-decouple==3.8
shortuuid==1.0.11
six==1.16.0
soupsieve==2.6
sqlparse==0.4.4
tzdata==2023.3
